feature/INS-784 - Change double to float

+++ ui-library/src/main/java/br/com/stant/libraries/uilibrary/components/simplepercentagebarview/SimplePercentageBarViewContract.java
@@ -6,6 +6,6 @@ package br.com.stant.libraries.uilibrary.components.simplepercentagebarview;
 
 public interface SimplePercentageBarViewContract {
 
-    void setSelectedPercent(float percentSelected);
-    void setExecutedPercent(float percentExecuted);
+    void setSelectedPercent(double percentSelected);
+    void setExecutedPercent(double percentExecuted);
 }
